split up inputs validation and errors tracking validation
place in a new folder 'Input/validation'
