+++ src/python/impactx/dashboard/Input/__init__.py
@@ -1,6 +1,6 @@
 from .defaults import DashboardDefaults
 from .utils import GeneralFunctions
-from .validation import DashboardValidation
+from .validation.inputs import DashboardValidation
 
 __all__ = [
     "DashboardDefaults",
