+++ src/python/impactx/dashboard/Input/validation/__init__.py
@@ -1,0 +1,7 @@
+from .inputs import DashboardValidation
+from .errors_tracker import sim_validator
+
+__all__ = [
+    "DashboardValidation",
+    "sim_validator",
+]
